adding appointments v1

+++ src/main/java/Hospital/Management/System/REST/API/entity/Appointments.java
@@ -16,7 +16,7 @@ import java.time.LocalDateTime;
 public class Appointments {
     @Id
     @GeneratedValue
-    private int AppointmentId;
+    private Long AppointmentId;
     private LocalDateTime Date;
     private Time StartTime;
     private Time EndTime;
